Fix test errors

+++ adabas/read_request_scan_test.go
@@ -128,7 +128,7 @@ func TestRequestLogicalWithQueryFieldsScan2(t *testing.T) {
 	assert.NoError(t, err)
 	var id int
 	var x, y, z string
-	m := 5.0
+	m := float32(5.0)
 	err = request.Scan(&id, &x, &y, &z, nil, &m)
 	assert.NoError(t, err)
